Make sb-vision an optional import
sb-vision is a pain to install for some due to opencv versions.

To much improve ease of development, make the dependency optional.
